+++ assets/blocks/djot/index.asset.php
@@ -9,5 +9,5 @@ return [
         'wp-i18n',
         'wp-api-fetch',
     ],
-    'version' => '1.5.10',
+    'version' => '1.5.11',
 ];